Style second section with css
The problem I had with this section is making the text be below the
image. I solved it by creating a max width for the container
of the image and text so that the text will automatically
move below the image.

Also, the images weren't all the same size, I solved it by using
object-fit: fill; that actually scales down the image and fills
the container, because if I used object-fit: scale-down; there will
be problems regarding if I used border (there will be white space
in either side of the image)

PS. border-radius makes borders round.